+++ src/components/templates/panel/PanelRight.tsx
@@ -1,11 +1,12 @@
 import dayjs from 'dayjs';
 
-import TodayMemo from '../todayMemo/TodayMemo';
-import TodayWork from '../todayWork/TodayWork';
+import TodayMemo from '~/components/organisms/todayMemo/TodayMemo';
+import TodayWork from '~/components/organisms/todayWork/TodayWork';
 
 import type { ICommonProps } from '~/types';
 
-const PanelRightMobile = ({ targetDate, userAgent }: ICommonProps) => {
+const PanelRight = ({ targetDate, userAgent }: ICommonProps) => {
+  const isMobile = userAgent === 'mobile';
   const date = dayjs(targetDate);
 
   return (
@@ -14,21 +15,18 @@ const PanelRightMobile = ({ targetDate, userAgent }: ICommonProps) => {
         backgroundColor: 'rgb(242, 242, 242)',
         flex: 1,
         width: '100%',
-        overflowY: 'scroll',
         padding: 0,
         height: 'calc(100vh - 50px)',
         display: 'flex',
         flexDirection: 'column',
       }}
     >
-      {/* <Header /> */}
       <div className='today-container' css={{ display: 'flex', height: '80px', alignItems: 'center' }}>
         <div css={{ fontSize: 20, fontWeight: 600, padding: 20 }}>
           {`${date.get('year')}년 ${date.get('month') + 1}월 ${date.get('date')}일의 워크로그`}
         </div>
       </div>
-
-      <div css={{ display: 'flex', flexDirection: 'column' }}>
+      <div css={{ height: 'calc(100% - 80px)', display: 'flex', flexDirection: isMobile ? 'column' : 'row' }}>
         <div className='today-task-container' css={{ flex: 0.5, padding: 20, borderRight: '1px solid #d5d5d552' }}>
           <TodayWork targetDate={targetDate} userAgent={userAgent} />
         </div>
@@ -40,4 +38,4 @@ const PanelRightMobile = ({ targetDate, userAgent }: ICommonProps) => {
   );
 };
 
-export default PanelRightMobile;
+export default PanelRight;
